Move main_liveTrade.py to project root
